requirements.txt
openai>=1.50.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0